+++ test/test-helpers.ts
@@ -20,6 +20,7 @@ import {
 } from '../src/outcome';
 import {State, hashState} from '../src/state';
 import {TransactionRequest} from 'ethers/providers';
+import {toHex} from '../src/hex-utils';
 
 export async function setupContracts(provider: ethers.providers.JsonRpcProvider, artifact) {
   const networkId = (await provider.getNetwork()).chainId;
@@ -55,14 +56,14 @@ export const ongoingChallengeHash = (turnNumRecord: number = 5) => {
 
 export const finalizedOutcomeHash = (
   turnNumRecord: number = 5,
-  finalizesAt: number = 1,
+  finalizesAt: string = toHex(1),
   challengerAddress: string = AddressZero,
   state?: State,
   outcome?: Outcome,
 ) => {
   return hashChannelStorage({
     largestTurnNum: bigNumberify(turnNumRecord).toHexString(),
-    finalizesAt: bigNumberify(finalizesAt).toHexString(),
+    finalizesAt,
     state,
     challengerAddress,
     outcome,
